+++ src/main/kotlin/joshxviii/plantz/entity/plant/Plant.kt
@@ -684,7 +684,7 @@ abstract class Plant(type: EntityType<out Plant>, level: Level) : TamableAnimal(
                 && itemStack.isEmpty
                 && player is ServerPlayer
                 && player.canWearPlant()
-                && player.isCrouching
+                && player.isShiftKeyDown
             ) {
                 if (!verifyOwner(player)) return InteractionResult.FAIL
                 if (attachToEntity(player)) {